The cuda now runs, but it produces an incorrect result.

+++ qt_aux/cuda1/MainWindow.h
@@ -48,7 +48,7 @@ private:
 
 
 private slots:
-    void updatesurf(double t, double *data);
+    void updatesurf(double t, float *data);
     void fpsUnbound();
 };
